refactor: Move to authentications directory.
